+++ src/MDBReader.ts
@@ -1,29 +1,32 @@
-import Database from "./Database.js";
-import PageType, { assertPageType } from "./PageType.js";
-import SysObject, { isSysObjectType, isSystemObject, SysObjectType } from "./SysObject.js";
-import Table from "./Table.js";
+import { Database } from "./Database.js";
+import { PageType, assertPageType } from "./PageType.js";
+import { SysObject, isSysObjectType, isSystemObject, SysObjectTypes } from "./SysObject.js";
+import { Table } from "./Table.js";
 import { SortOrder } from "./types.js";
 
 const MSYS_OBJECTS_TABLE = "MSysObjects";
 const MSYS_OBJECTS_PAGE = 2;
 
 export interface Options {
-    password?: string;
+    password?: string | undefined;
 }
 
 export default class MDBReader {
-    private readonly sysObjects: SysObject[];
-    private readonly db: Database;
+    #buffer: Buffer;
+    #sysObjects: SysObject[];
+    #database: Database;
 
     /**
      * @param buffer Buffer of the database.
      */
-    public constructor(private readonly buffer: Buffer, { password }: Options = {}) {
-        assertPageType(this.buffer, PageType.DatabaseDefinitionPage);
+    constructor(buffer: Buffer, { password }: Options | undefined = {}) {
+        this.#buffer = buffer;
 
-        this.db = new Database(this.buffer, password ?? "");
+        assertPageType(this.#buffer, PageType.DatabaseDefinitionPage);
 
-        const mSysObjectsTable = new Table(MSYS_OBJECTS_TABLE, this.db, MSYS_OBJECTS_PAGE).getData<{
+        this.#database = new Database(this.#buffer, password ?? "");
+
+        const mSysObjectsTable = new Table(MSYS_OBJECTS_TABLE, this.#database, MSYS_OBJECTS_PAGE).getData<{
             Id: number;
             Name: string;
             Type: number;
@@ -32,7 +35,7 @@ export default class MDBReader {
             columns: ["Id", "Name", "Type", "Flags"],
         });
 
-        this.sysObjects = mSysObjectsTable.map((mSysObject) => {
+        this.#sysObjects = mSysObjectsTable.map((mSysObject) => {
             const objectType = mSysObject.Type & 0x7f;
             return {
                 objectName: mSysObject.Name,
@@ -46,22 +49,22 @@ export default class MDBReader {
     /**
      * Date when the database was created
      */
-    public getCreationDate(): Date | null {
-        return this.db.getCreationDate();
+    getCreationDate(): Date | null {
+        return this.#database.getCreationDate();
     }
 
     /**
      * Database password
      */
-    public getPassword(): string | null {
-        return this.db.getPassword();
+    getPassword(): string | null {
+        return this.#database.getPassword();
     }
 
     /**
      * Default sort order
      */
-    public getDefaultSortOrder(): Readonly<SortOrder> {
-        return this.db.getDefaultSortOrder();
+    getDefaultSortOrder(): Readonly<SortOrder> {
+        return this.#database.getDefaultSortOrder();
     }
 
     /**
@@ -71,20 +74,18 @@ export default class MDBReader {
      * @param systemTables Includes system tables. Default false.
      * @param linkedTables Includes linked tables. Default false.
      */
-    public getTableNames(
-        {
-            normalTables,
-            systemTables,
-            linkedTables,
-        }: {
-            normalTables: boolean;
-            systemTables: boolean;
-            linkedTables: boolean;
-        } = { normalTables: true, systemTables: false, linkedTables: false }
-    ): string[] {
+    getTableNames({
+        normalTables = true,
+        systemTables = false,
+        linkedTables = false,
+    }: {
+        normalTables?: boolean | undefined;
+        systemTables?: boolean | undefined;
+        linkedTables?: boolean | undefined;
+    } = {}): string[] {
         const filteredSysObjects: SysObject[] = [];
-        for (const sysObject of this.sysObjects) {
-            if (sysObject.objectType === SysObjectType.Table) {
+        for (const sysObject of this.#sysObjects) {
+            if (sysObject.objectType === SysObjectTypes.Table) {
                 if (!isSystemObject(sysObject)) {
                     if (normalTables) {
                         filteredSysObjects.push(sysObject);
@@ -92,7 +93,7 @@ export default class MDBReader {
                 } else if (systemTables) {
                     filteredSysObjects.push(sysObject);
                 }
-            } else if (sysObject.objectType === SysObjectType.LinkedTable && linkedTables) {
+            } else if (sysObject.objectType === SysObjectTypes.LinkedTable && linkedTables) {
                 filteredSysObjects.push(sysObject);
             }
         }
@@ -105,15 +106,15 @@ export default class MDBReader {
      *
      * @param name Name of the table. Case sensitive.
      */
-    public getTable(name: string): Table {
-        const sysObject = this.sysObjects
-            .filter((o) => o.objectType === SysObjectType.Table)
+    getTable(name: string): Table {
+        const sysObject = this.#sysObjects
+            .filter((o) => o.objectType === SysObjectTypes.Table)
             .find((o) => o.objectName === name);
 
         if (!sysObject) {
             throw new Error(`Could not find table with name ${name}`);
         }
 
-        return new Table(name, this.db, sysObject.tablePage);
+        return new Table(name, this.#database, sysObject.tablePage);
     }
 }