+++ src/PageType.ts
@@ -1,7 +1,7 @@
 /**
  * @see https://github.com/brianb/mdbtools/blob/d6f5745d949f37db969d5f424e69b54f0da60b9b/HACKING#L64-L70
  */
-enum PageType {
+export const enum PageType {
     DatabaseDefinitionPage = 0x00,
     DataPage = 0x01,
     TableDefinition = 0x02,
@@ -9,8 +9,6 @@ enum PageType {
     LeafIndexPages = 0x04,
     PageUsageBitmaps = 0x05,
 }
-
-export default PageType;
 
 export function assertPageType(buffer: Buffer, pageType: PageType): void {
     if (buffer[0] !== pageType) {