+++ src/crypto/deriveKey.ts
@@ -2,7 +2,7 @@ import { fixBufferLength, intToBuffer } from "../util.js";
 import { hash } from "./hash.js";
 
 /**
- * Can probably be replaced with `crypto.webcrypto.subtle.derivekey(...)` once node 14 support is dropped
+ * Can probably be replaced with `crypto.webcrypto.subtle.derivekey(...)` once node 16 support is dropped
  */
 export function deriveKey(
     password: Buffer,