+++ src/index.ts
@@ -1,4 +1,4 @@
 export type { Column } from "./column.js";
-export { default, Options } from "./MDBReader.js";
-export type { default as Table } from "./Table.js";
+export { default, type Options } from "./MDBReader.js";
+export type { Table } from "./Table.js";
 export * from "./types.js";